pyproject.toml
[project]
name = "spam-classification"
version = "0.1.0"
description = "SMS Spam Classification with ML and Agent"
requires-python = ">=3.12"
dependencies = [
    "polars>=0.20.0",
    "pandas>=2.2.0",
    "pydantic>=2.0.0",
    "pandera>=0.20.0",
    "scikit-learn>=1.5.0",
    "lightgbm>=4.0.0",
    "nltk>=3.8.0",
    "pydantic-ai>=0.0.0",
    "openai>=1.0.0",
    "seaborn>=0.13.0",
    "streamlit>=1.28.0",
    "python-dotenv>=1.0.0",
    "numpy>=1.24.0",
    "tqdm>=4.65.0",
    "plotly>=5.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.1.0",
    "black>=23.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = []

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.black]
line-length = 100
target-version = ['py312']
